requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
markdown>=3.5.0
python-frontmatter>=1.1.0
Pygments>=2.17.0
httpx>=0.27.0
